Grouped color formats under the Color class and changed the to_hue method to to_h
