+++ lib/hue.rb
@@ -3,9 +3,9 @@ require 'hue/errors'
 require 'hue/client'
 require 'hue/bridge'
 require 'hue/light'
-require 'hue/rgb'
-require 'hue/hsl'
-require 'hue/hexrgb'
+require 'hue/color/rgb'
+require 'hue/color/hsl'
+require 'hue/color/hexrgb'
 
 module Hue
   USERNAME_RANGE = 10..40
