--- lib/hue/hexrgb.rb
@@ -1,8 +1,0 @@
-module Hue
-  class HexRGB < RGB
-    def initialize(hex)
-      hex = hex.scan(/../).map { |e| e.to_i(16) }
-      super(hex[0], hex[1], hex[2])
-    end
-  end
-end
